[core] Cleanup Xposed Bridge implementation
