+++ core/src/main/java/de/robv/android/xposed/PendingHooks.java
@@ -24,9 +24,9 @@ import java.lang.reflect.Member;
 import java.lang.reflect.Method;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.function.Function;
 
-import static de.robv.android.xposed.XposedBridge.hookMethodNative;
+import io.github.lsposed.lspd.config.LSPdConfigGlobal;
+
 import static io.github.lsposed.lspd.nativebridge.PendingHooks.recordPendingMethodNative;
 
 public final class PendingHooks {
@@ -38,7 +38,7 @@ public final class PendingHooks {
     public synchronized static void hookPendingMethod(Class<?> clazz) {
         if (sPendingHooks.containsKey(clazz)) {
             for (Map.Entry<Member, XposedBridge.AdditionalHookInfo> hook : sPendingHooks.get(clazz).entrySet()) {
-                hookMethodNative(hook.getKey(), clazz, 0, hook.getValue());
+                LSPdConfigGlobal.getHookProvider().hookMethod(hook.getKey(), hook.getValue());
             }
             sPendingHooks.remove(clazz);
         }
@@ -47,13 +47,7 @@ public final class PendingHooks {
     public synchronized static void recordPendingMethod(Method hookMethod,
                                                         XposedBridge.AdditionalHookInfo additionalInfo) {
         ConcurrentHashMap<Member, XposedBridge.AdditionalHookInfo> pending =
-                sPendingHooks.computeIfAbsent(hookMethod.getDeclaringClass(),
-                        new Function<Class<?>, ConcurrentHashMap<Member, XposedBridge.AdditionalHookInfo>>() {
-                            @Override
-                            public ConcurrentHashMap<Member, XposedBridge.AdditionalHookInfo> apply(Class<?> aClass) {
-                                return new ConcurrentHashMap<>();
-                            }
-                        });
+                sPendingHooks.computeIfAbsent(hookMethod.getDeclaringClass(), aClass -> new ConcurrentHashMap<>());
 
         pending.put(hookMethod, additionalInfo);
         recordPendingMethodNative(hookMethod, hookMethod.getDeclaringClass());
