+++ core/src/main/java/de/robv/android/xposed/SELinuxHelper.java
@@ -20,12 +20,6 @@
 
 package de.robv.android.xposed;
 
-import android.os.SELinux;
-
-import java.io.File;
-import java.io.FileInputStream;
-import java.io.IOException;
-
 import de.robv.android.xposed.services.BaseService;
 import de.robv.android.xposed.services.DirectAccessService;
 
@@ -41,7 +35,8 @@ public final class SELinuxHelper {
 	 * @return A boolean indicating whether SELinux is enabled.
 	 */
 	public static boolean isSELinuxEnabled() {
-		return sIsSELinuxEnabled;
+		// lsp: always enabled
+		return true;
 	}
 
 	/**
@@ -50,36 +45,8 @@ public final class SELinuxHelper {
 	 * @return A boolean indicating whether SELinux is enforcing.
 	*/
 	public static boolean isSELinuxEnforced() {
-		if (!sIsSELinuxEnabled) {
-			return false;
-		}
-		boolean result = false;
-		final File SELINUX_STATUS_FILE = new File("/sys/fs/selinux/enforce");
-		if (SELINUX_STATUS_FILE.exists()) {
-			try {
-				FileInputStream fis = new FileInputStream(SELINUX_STATUS_FILE);
-				int status = fis.read();
-				switch (status) {
-					case 49:
-						result = true;
-						break;
-					case 48:
-						result = false;
-						break;
-					default:
-						XposedBridge.log("Unexpected byte " + status + " in /sys/fs/selinux/enforce");
-				}
-				fis.close();
-			} catch (IOException e) {
-				if (e.getMessage().contains("Permission denied")) {
-					result = true;
-				} else {
-					XposedBridge.log("Failed to read SELinux status: " + e.getMessage());
-					result = false;
-				}
-			}
-		}
-		return result;
+		// lsp: always enforcing
+		return true;
 	}
 
 	/**
@@ -88,7 +55,7 @@ public final class SELinuxHelper {
 	 * @return A String representing the security context of the current process.
 	 */
 	public static String getContext() {
-		return sIsSELinuxEnabled ? SELinux.getContext() : null;
+		return null;
 	}
 
 	/**
@@ -100,36 +67,9 @@ public final class SELinuxHelper {
 	 * @return An instance of the service.
 	 */
 	public static BaseService getAppDataFileService() {
-		if (sServiceAppDataFile != null)
-			return sServiceAppDataFile;
-		throw new UnsupportedOperationException();
-	}
+        return sServiceAppDataFile;
+    }
 
+	private static final BaseService sServiceAppDataFile = new DirectAccessService(); // ed: initialized directly
 
-	// ----------------------------------------------------------------------------
-	// TODO: SELinux status
-	private static boolean sIsSELinuxEnabled = false;
-	private static BaseService sServiceAppDataFile = new DirectAccessService(); // ed: initialized directly
-
-	/*package*/ public static void initOnce() {
-		// ed: we assume all selinux policies have been added lively using magiskpolicy
-		try {
-			sIsSELinuxEnabled = SELinux.isSELinuxEnabled();
-		} catch (NoClassDefFoundError ignored) {}
-	}
-
-	/*package*/ static void initForProcess(String packageName) {
-		// ed: sServiceAppDataFile has been initialized with default value
-//		if (sIsSELinuxEnabled) {
-//			if (packageName == null) {  // Zygote
-//				sServiceAppDataFile = new ZygoteService();
-//			} else if (packageName.equals("android")) {  //system_server
-//				sServiceAppDataFile = BinderService.getService(BinderService.TARGET_APP);
-//			} else {  // app
-//				sServiceAppDataFile = new DirectAccessService();
-//			}
-//		} else {
-//			sServiceAppDataFile = new DirectAccessService();
-//		}
-	}
 }
