+++ core/src/main/java/de/robv/android/xposed/XposedHelpers.java
@@ -1645,70 +1645,6 @@ public final class XposedHelpers {
 		}
 	}
 
-	/*package*/ static boolean fileContains(File file, String str) throws IOException {
-		// There are certainly more efficient algorithms (e.g. Boyer-Moore used in grep),
-		// but the naive approach should be sufficient here.
-		BufferedReader in = null;
-		try {
-			in = new BufferedReader(new FileReader(file));
-			String line;
-			while ((line = in.readLine()) != null) {
-				if (line.contains(str)) {
-					return true;
-				}
-			}
-			return false;
-		} finally {
-			closeSilently(in);
-		}
-	}
-
-	//#################################################################################################
-
-	/**
-	 * Returns the method that is overridden by the given method.
-	 * It returns {@code null} if the method doesn't override another method or if that method is
-	 * abstract, i.e. if this is the first implementation in the hierarchy.
-	 */
-	/*package*/ static Method getOverriddenMethod(Method method) {
-		int modifiers = method.getModifiers();
-		if (Modifier.isStatic(modifiers) || Modifier.isPrivate(modifiers)) {
-			return null;
-		}
-
-		String name = method.getName();
-		Class<?>[] parameters = method.getParameterTypes();
-		Class<?> clazz = method.getDeclaringClass().getSuperclass();
-		while (clazz != null) {
-			try {
-				Method superMethod = clazz.getDeclaredMethod(name, parameters);
-				modifiers = superMethod.getModifiers();
-				if (!Modifier.isPrivate(modifiers) && !Modifier.isAbstract(modifiers)) {
-					return superMethod;
-				} else {
-					return null;
-				}
-			} catch (NoSuchMethodException ignored) {
-				clazz = clazz.getSuperclass();
-			}
-		}
-		return null;
-	}
-
-	/**
-	 * Returns all methods which this class overrides.
-	 */
-	/*package*/ static Set<Method> getOverriddenMethods(Class<?> clazz) {
-		Set<Method> methods = new HashSet<>();
-		for (Method method : clazz.getDeclaredMethods()) {
-			Method overridden = getOverriddenMethod(method);
-			if (overridden != null) {
-				methods.add(overridden);
-			}
-		}
-		return methods;
-	}
-
 	//#################################################################################################
 	// TODO helpers for view traversing
 	/*To make it easier, I will try and implement some more helpers:
