+++ core/src/main/java/io/github/lsposed/lspd/annotation/Level.java
@@ -18,20 +18,12 @@
  * Copyright (C) 2021 LSPosed Contributors
  */
 
-package de.robv.android.xposed.annotation;
+package io.github.lsposed.lspd.annotation;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+public enum Level {
+    LOW, MIDDLE, HIGH;
 
-/**
- * Annotation that indicates a element is sensitive to Android API level.
- * <p>
- * Annotated elements' compatibility should be checked when adapting to new Android versions.
- */
-@Retention(RetentionPolicy.SOURCE)
-@Target({ElementType.METHOD, ElementType.FIELD, ElementType.TYPE})
-public @interface ApiSensitive {
-    Level value() default Level.HIGH;
+    private Level() {
+
+    }
 }
