+++ core/src/main/java/io/github/lsposed/lspd/proxy/NormalProxy.java
@@ -28,8 +28,6 @@ import io.github.lsposed.lspd.deopt.PrebuiltMethodsDeopter;
 import io.github.lsposed.lspd.nativebridge.ModuleLogger;
 import io.github.lsposed.lspd.util.Utils;
 
-import de.robv.android.xposed.SELinuxHelper;
-
 import static io.github.lsposed.lspd.config.LSPApplicationServiceClient.serviceClient;
 
 public class NormalProxy extends BaseProxy {
@@ -51,7 +49,6 @@ public class NormalProxy extends BaseProxy {
     private void forkPostCommon(boolean isSystem, String appDataDir, String niceName) {
         // init logger
         ModuleLogger.initLogger(serviceClient.getModuleLogger());
-        SELinuxHelper.initOnce();
         mRouter.initResourcesHook();
         mRouter.prepare(isSystem);
         PrebuiltMethodsDeopter.deoptBootMethods(); // do it once for secondary zygote
