+++ core/src/main/java/io/github/lsposed/lspd/util/ClassLoaderUtils.java
@@ -35,8 +35,8 @@ import dalvik.system.PathClassLoader;
 import de.robv.android.xposed.XC_MethodHook;
 import de.robv.android.xposed.XposedBridge;
 import de.robv.android.xposed.XposedHelpers;
-import de.robv.android.xposed.annotation.ApiSensitive;
-import de.robv.android.xposed.annotation.Level;
+import io.github.lsposed.lspd.annotation.ApiSensitive;
+import io.github.lsposed.lspd.annotation.Level;
 
 @ApiSensitive(Level.LOW)
 public class ClassLoaderUtils {
