+++ core/src/main/java/io/github/lsposed/lspd/util/ProcessUtils.java
@@ -28,8 +28,8 @@ import java.io.File;
 import java.io.FileInputStream;
 import java.io.InputStreamReader;
 
-import de.robv.android.xposed.annotation.ApiSensitive;
-import de.robv.android.xposed.annotation.Level;
+import io.github.lsposed.lspd.annotation.ApiSensitive;
+import io.github.lsposed.lspd.annotation.Level;
 
 @ApiSensitive(Level.LOW)
 public class ProcessUtils {
