+++ core/src/main/java/io/github/lsposed/lspd/util/Unsafe.java
@@ -25,8 +25,8 @@ import android.util.Log;
 import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 
-import de.robv.android.xposed.annotation.ApiSensitive;
-import de.robv.android.xposed.annotation.Level;
+import io.github.lsposed.lspd.annotation.ApiSensitive;
+import io.github.lsposed.lspd.annotation.Level;
 
 @ApiSensitive(Level.LOW)
 public final class Unsafe {
