+++ core/src/main/java/io/github/lsposed/lspd/util/Utils.java
@@ -25,8 +25,8 @@ import android.util.Log;
 import io.github.lsposed.lspd.BuildConfig;
 
 import de.robv.android.xposed.XposedHelpers;
-import de.robv.android.xposed.annotation.ApiSensitive;
-import de.robv.android.xposed.annotation.Level;
+import io.github.lsposed.lspd.annotation.ApiSensitive;
+import io.github.lsposed.lspd.annotation.Level;
 
 
 public class Utils {
